+++ src/dictionary/DefaultEditButtonGroup.ts
@@ -5,6 +5,7 @@ import { renderType } from "./DefaultMod"
 
 export interface DefaultEditButtonGroupOption extends DefaultEditButtonOption {
   disabled?: boolean
+  name: string
   render?: renderType
 }